Add test rules for `Google Pay` & fraud signals
